requirements.txt
# Core LLM & Agent Framework
langchain>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.2.0
langgraph>=0.2.0

# Anthropic SDK (underlying)
anthropic>=0.34.0

# Utilities
python-dotenv>=1.0.0

# Optional: for richer terminal output
rich>=13.0.0

# Web server (for frontend)
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
